chore: move tests to test folder
